+++ src/cpp_common/utilities.cpp
@@ -123,6 +123,8 @@ get_name(Which which) {
         case STRONGCOMPONENTS:
             return "pgr_strongComponents";
             break;
+        case PLANARFACES:
+            return "pgr_planarFaces";
         case CORENUMBERS:
             return "pgr_coreNumbers";
             break;