+++ NEWS.md
@@ -38,6 +38,10 @@ To see all issues & pull requests closed by this release see the
 
 **Summary of changes by function**
 
+* pgr_planarFaces
+
+  * New experimental function.
+
 * pgr_coreNumbers
 
   * New experimental function.