+++ doc/src/release_notes.rst
@@ -63,6 +63,12 @@ To see all issues & pull requests closed by this release see the
 
 .. rubric:: Summary of changes by function
 
+* pgr_planarFaces
+
+  .. include:: pgr_planarFaces.rst
+     :start-after: Version 4.1.0
+     :end-before: Description
+
 * pgr_coreNumbers
 
   .. include:: pgr_coreNumbers.rst